added utils/email to handle sending emails
