+++ views/partial/_emailPasswordResetRequest.ejs
@@ -6,7 +6,7 @@
     <body>
         <p>Hi <%= name %>,</p>
         <p>You requested to reset your password.</p>
-        <p> Please, click the link below to reset your password</p>
+        <p> Please, click the following link to reset your passwor:</p>
         <a href="https://<%= link %>">Reset Password</a>
     </body>
 </html>
